Restructured samples and tests.
All samples should now be located in their own samples/sampleName directory. All samples are built simultenously with the samples.hxml build file.

Tests are now located in the tests/ directory. The tests.hxml build file builds a Neko executable in the tests/bin directory, which is used to run the test cases.