(Started from ground up. Introduced basic Java and JavaScript render targets.)
The rewrite is focused on render target creation before pipeline implementation. The proposed target creation process includes three classes: `Context`, `FrameBuffer`, and `FrameBufferData`.

`Context` must handle target window creation where the render, as represented by the frame buffer, is drawn to. This makes the `draw()` method (and signature) a target requirement.

`FrameBuffer` must work with an underlying data implementation to provide methods for getting and setting individual pixels and their channels, as well as buffer clearing methods.

`FrameBufferData` is usually a typedef that signifies the data structure or type holding frame buffer pixel data.

The Java target is intended for desktop use and requires `hxjava` as a dependency. It uses AWT and Swing to handle windowing, and the frame buffer is represented by an array of 32-bit integers where every integer contains data for a single pixel.

The JavaScript target is intended for web use and relies on the HTML5 canvas element. The frame buffer is represented by an array of clamped unsigned 8-bit integers where every integer contains data for a colour channel of a single pixel.