Make vectors, matrices abstracts over custom types
The previous implementation of vectors and matrices as abstracts over
integer and float arrays proved itself to be flawed. It caused significant
allocation and memory leakage on the Java target, which in turn made the
renderer stutter and perform slowly.

Custom vector and matrix types reduce the overhead of an array type as
they only serve as a collection of vector or matrix elements, with the
addition ofa `toString()` method for easier debugging.

Inline constructors are now also possible, which should result in a
significant performance increase.