Implement rudimentary clipping
Clipping is performed in two passes. The first pass is performed on
vertices during the vertex post-processing stage. At this stage,
primitives are clipped (discarded) if they're not within the view volume.

The second stage is performed during scan conversion. When a projected
primitive's rasterization area is determined, it is limited to the
framebuffer's dimensions.